requirements.txt
# GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.1

# OCR для капчи
easyocr>=1.7.0
Pillow>=10.1.0
numpy>=1.24.0

# Excel обработка
openpyxl>=3.1.2

# Дополнительно
requests>=2.31.0

# GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.1
selenium-stealth>=1.0.6
undetected-chromedriver>=3.5.4

# OCR для капчи
easyocr>=1.7.0
Pillow>=10.1.0
numpy>=1.24.0

# Excel обработка
openpyxl>=3.1.2

# Telegram Bot
python-telegram-bot>=20.0

# System & Security
psutil>=5.9.0
cryptography>=41.0.0

# Дополнительно
requests>=2.31.0